+++ src/renderer/pages/author/author.register.page.tsx
@@ -6,14 +6,14 @@ import {
 	useActionData,
 	useLoaderData,
 } from "react-router-dom";
-import { AuthorRegisterTemplate } from "../../../features/author/ui/templates/author.register.template.js";
-import { ApiService } from "../autogenerate/register.js";
+import { AuthorRegisterTemplate } from "../../../../features/author/ui/templates/author.register.template.js";
+import { ApiService } from "../../autogenerate/register.js";
 import {
 	type ActionError,
 	getErrorMessage,
 	isActionError,
 	isErrorResponse,
-} from "../utils/error.js";
+} from "../../utils/error.js";
 
 const client = new ApiService();
