+++ src/renderer/pages/illust/illust.edit.page.tsx
@@ -7,47 +7,52 @@ import {
 	useLoaderData,
 	useNavigate,
 } from "react-router-dom";
-import { AuthorEditTemplate } from "../../../features/author/ui/templates/author.edit.template.js";
-import { ApiService } from "../autogenerate/register.js";
+import type { Illust } from "../../../../features/illust/illust.model.js";
+import { IllustEditTemplate } from "../../../../features/illust/ui/templates/illust.edit.template.js";
+import { ApiService } from "../../autogenerate/register.js";
 import {
 	type ActionError,
 	getErrorMessage,
 	isActionError,
 	isErrorResponse,
-} from "../utils/error.js";
+} from "../../utils/error.js";
 
 const client = new ApiService();
 
 export async function loader({ params }: LoaderFunctionArgs) {
-	const { authorId } = params;
-	if (!authorId) {
-		throw new Error("Author ID is required");
+	const { illustId } = params;
+	if (!illustId) {
+		throw new Error("Illust ID is required");
 	}
 
-	const response = await client.getAuthorDetail(authorId, 1, 20);
+	const response = await client.detailIllust(illustId);
 	if (isFailure(response)) {
 		throw response.value;
 	}
-	return {
-		id: response.value.id,
-		name: response.value.name,
-		urls: response.value.urls,
-	};
+	return response.value;
 }
 
 export async function action({ request, params }: ActionFunctionArgs) {
-	const { authorId } = params;
-	if (!authorId) {
-		throw new Error("Author ID is required");
+	const { illustId } = params;
+	if (!illustId) {
+		throw new Error("Illust ID is required");
 	}
 
 	const formData = await request.formData();
-	const name = formData.get("name")?.toString() ?? "";
-	const urls = formData.get("urls")?.toString() || "{}";
+	const tags = formData.get("tags")?.toString() ?? "";
+	const imageItems = formData.getAll("imageItems").map(String);
+	const authorIds = formData.getAll("authorIds").map(String);
 
-	console.log(`id: ${authorId}, name: ${name}, urls: ${urls}`);
+	console.log(
+		`url: ${request.url}, id: ${illustId}, tags: ${tags}, imageItems: ${imageItems}, authorIds: ${authorIds}`,
+	);
 
-	const response = await client.updateAuthor(authorId, name, urls);
+	const response = await client.updateIllust(
+		illustId,
+		tags,
+		imageItems,
+		authorIds,
+	);
 	if (isFailure(response)) {
 		const error = response.value;
 		console.log(`response error: ${getErrorMessage(error)}`);
@@ -57,16 +62,12 @@ export async function action({ request, params }: ActionFunctionArgs) {
 		throw new Error(getErrorMessage(error));
 	}
 
-	// 更新成功後、作者詳細画面へリダイレクト
-	return redirect(`/author/${authorId}`);
+	// 更新成功後、詳細画面へリダイレクト
+	return redirect(`/illust/${illustId}`);
 }
 
-export default function AuthorEditPage() {
-	const author = useLoaderData<{
-		id: string;
-		name: string;
-		urls: Record<string, string>;
-	}>();
+export default function IllustEditPage() {
+	const illust = useLoaderData<Illust>();
 	const actionData = useActionData<ActionError>();
 	const navigate = useNavigate();
 
@@ -82,12 +83,12 @@ export default function AuthorEditPage() {
 			: undefined;
 
 	const handleCancel = () => {
-		navigate(`/author/${author.id}`);
+		navigate(`/illust/${illust.id}`);
 	};
 
 	return (
-		<AuthorEditTemplate
-			author={author}
+		<IllustEditTemplate
+			illust={illust}
 			onCancel={handleCancel}
 			fieldErrors={fieldErrors}
 			generalError={generalError}
