+++ src/renderer/route.tsx
@@ -7,18 +7,18 @@ import AppSettingPage, {
 } from "./pages/app.setting.page.js";
 import AuthorDetailPage, {
 	loader as authorDetailLoader,
-} from "./pages/author.detail.page.js";
+} from "./pages/author/author.detail.page.js";
 import AuthorEditPage, {
 	action as authorEditAction,
 	loader as authorEditLoader,
-} from "./pages/author.edit.page.js";
+} from "./pages/author/author.edit.page.js";
 import AuthorRegisterPage, {
 	action as authorRegisterAction,
 	loader as authorRegisterLoader,
-} from "./pages/author.register.page.js";
+} from "./pages/author/author.register.page.js";
 import AuthorSearchPage, {
 	loader as authorSearchLoader,
-} from "./pages/author.search.page.js";
+} from "./pages/author/author.search.page.js";
 import DuplicateDetailPage, {
 	loader as duplicateDetailLoader,
 } from "./pages/duplicate.detail.page.js";
@@ -27,28 +27,28 @@ import DuplicateListPage, {
 } from "./pages/duplicate.list.page.js";
 import IllustDetailPage, {
 	loader as illustDetailLoader,
-} from "./pages/illust.detail.page.js";
+} from "./pages/illust/illust.detail.page.js";
 import IllustEditPage, {
 	action as illustEditAction,
 	loader as illustEditLoader,
-} from "./pages/illust.edit.page.js";
+} from "./pages/illust/illust.edit.page.js";
 import IllustRegisterPage, {
 	action as illustRegisterAction,
 	loader as illustRegisterLoader,
-} from "./pages/illust.register.page.js";
+} from "./pages/illust/illust.register.page.js";
 import IllustSearchPage, {
 	loader as illustSearchLoader,
-} from "./pages/illust.search.page.js";
+} from "./pages/illust/illust.search.page.js";
 import VideoDetailPage, {
 	loader as videoDetailLoader,
-} from "./pages/video.detail.page.js";
+} from "./pages/video/video.detail.page.js";
 import VideoRegisterPage, {
 	action as videoRegisterAction,
 	loader as videoRegisterLoader,
-} from "./pages/video.register.page.js";
+} from "./pages/video/video.register.page.js";
 import VideoSearchPage, {
 	loader as videoSearchLoader,
-} from "./pages/video.search.page.js";
+} from "./pages/video/video.search.page.js";
 
 export const route = createHashRouter([
 	{
